# backend/docs/swagger.yaml
definitions:
  github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError:
    properties:
      code:
        enum:
        - INVALID_REQUEST
        - ' INTERNAL_ERROR'
        - ' SCHEDULE_EXISTS'
        - ' SLOT_ALREADY_BOOKED'
        - ' FORBIDEN'
        - ' NOT_FOUND'
        type: string
      message:
        example: invalid request
        type: string
    type: object
  github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError:
    properties:
      code:
        enum:
        - INTERNAL_ERROR
        type: string
      message:
        example: internal server error
        type: string
    type: object
  github_com_tryingmyb3st_PolyTweet_internal_core_domain.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_tryingmyb3st_PolyTweet_internal_core_domain.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      parent_id:
        type: string
      reply_to:
        type: string
      user_id:
        type: string
    type: object
  internal_features_auth_transport_http.DummyLoginDTO:
    properties:
      role:
        example: user
        type: string
    type: object
  internal_features_auth_transport_http.DummyLoginDTOResponse:
    properties:
      token:
        type: string
    type: object
  internal_features_auth_transport_http.LoginDTO:
    properties:
      email:
        example: useremail@gmail.com
        type: string
      password:
        example: supersecretpass
        type: string
    type: object
  internal_features_auth_transport_http.LoginDTOResponse:
    properties:
      token:
        type: string
    type: object
  internal_features_auth_transport_http.RegisterDTO:
    properties:
      email:
        example: useremail@gmail.com
        type: string
        x-order: "0"
      password:
        example: supersecretpass
        type: string
        x-order: "1"
      role:
        example: admin
        type: string
        x-order: "2"
    type: object
  internal_features_auth_transport_http.RegisterDTOResponse:
    properties:
      createdAt:
        example: "2026-03-25T12:00:41.267Z"
        type: string
        x-order: "4"
      email:
        example: useremail@gmail.com
        type: string
        x-order: "1"
      id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
        x-order: "0"
      password:
        type: string
        x-order: "2"
      role:
        example: admin
        type: string
        x-order: "3"
    type: object
  internal_features_posts_transport_http.CreatePostDTO:
    properties:
      content:
        example: Hello, world!
        type: string
        x-order: "0"
      image_url:
        example: https://example.com/image.jpg
        type: string
        x-order: "3"
      parent_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
        x-order: "1"
      reply_to:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
        x-order: "2"
    type: object
  internal_features_posts_transport_http.CreatePostDTOResponse:
    properties:
      content:
        example: Hello, world!
        type: string
        x-order: "2"
      created_at:
        example: "2026-03-25T12:00:41.267Z"
        type: string
        x-order: "6"
      id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
        x-order: "0"
      image_url:
        example: https://example.com/image.jpg
        type: string
        x-order: "5"
      parent_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
        x-order: "3"
      reply_to:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
        x-order: "4"
      user_id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
        x-order: "1"
    type: object
  internal_features_posts_transport_http.DeletePostDTOResponse:
    properties:
      message:
        type: string
      post_id:
        type: string
    type: object
  internal_features_posts_transport_http.GetLastWeekPostsDTOResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.Pagination'
      posts:
        items:
          $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.Post'
        type: array
    type: object
  internal_features_posts_transport_http.GetPostByIdDTOResponse:
    properties:
      content:
        example: Hello, world!
        type: string
        x-order: "2"
      created_at:
        example: "2026-03-25T12:00:41.267Z"
        type: string
        x-order: "6"
      id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
        x-order: "0"
      image_url:
        example: https://example.com/image.jpg
        type: string
        x-order: "5"
      parent_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
        x-order: "3"
      reply_to:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
        x-order: "4"
      user_id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
        x-order: "1"
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  title: Poly Twitter Service
  version: "1.0"
paths:
  /dummyLogin:
    post:
      consumes:
      - application/json
      description: 'Выдаёт тестовый JWT для указанной роли (admin / user). Для каждой
        роли возвращается фиксированный UUID пользователя: один и тот же UUID для
        всех запросов с ролью admin и один и тот же UUID для роли user.'
      parameters:
      - description: 'доступны роли: admin, user'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_auth_transport_http.DummyLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Тестовый токен
          schema:
            $ref: '#/definitions/internal_features_auth_transport_http.DummyLoginDTOResponse'
        "400":
          description: Неверный запрос (недопустимое значение роли)
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Получить тестовый JWT по роли
      tags:
      - Auth
  /login:
    post:
      consumes:
      - application/json
      description: Авторизует пользователя по email и паролю, возвращает JWT. Для
        авторизации в рамках теста - /dummyLogin.
      parameters:
      - description: тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_auth_transport_http.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация
          schema:
            $ref: '#/definitions/internal_features_auth_transport_http.LoginDTOResponse'
        "401":
          description: Неверные учётные данные
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Авторизация по email и паролю
      tags:
      - Auth
  /posts/{PostId}:
    get:
      consumes:
      - application/json
      description: Ищет пост по ID
      parameters:
      - description: Post ID
        in: path
        name: PostId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пост найден
          schema:
            $ref: '#/definitions/internal_features_posts_transport_http.GetPostByIdDTOResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "404":
          description: Пост не найден
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Поиск поста по ID
      tags:
      - Posts
  /posts/{PostId}/delete:
    delete:
      consumes:
      - application/json
      description: Удаляет пост, если запрос на удаление исходит от автора поста
      parameters:
      - description: Post ID (UUID)
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        format: uuid
        in: path
        name: postId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Пост успешно удален
          schema:
            $ref: '#/definitions/internal_features_posts_transport_http.DeletePostDTOResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "404":
          description: Пост не найден
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Удаление поста
      tags:
      - Posts
  /posts/all:
    get:
      consumes:
      - application/json
      description: Ищет посты за последние 7 дней с поддержкой пагинации через параметры
        page и pageSize.
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 15
        description: Размер страницы
        in: query
        maximum: 30
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Посты найдены
          schema:
            $ref: '#/definitions/internal_features_posts_transport_http.GetLastWeekPostsDTOResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Поиск постов за последние 7 дней
      tags:
      - Posts
  /posts/create:
    post:
      consumes:
      - application/json
      description: Создает пост длиной <= 280 слов
      parameters:
      - description: тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_posts_transport_http.CreatePostDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Пост создан
          schema:
            $ref: '#/definitions/internal_features_posts_transport_http.CreatePostDTOResponse'
        "400":
          description: Неверный запрос или content > 280
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Создание поста
      tags:
      - Posts
  /posts/users/{UserId}:
    get:
      consumes:
      - application/json
      description: Ищет посты пользователя по ID с поддержкой пагинации через параметры
        page и pageSize.
      parameters:
      - description: User ID
        in: path
        name: UserId
        required: true
        type: integer
      - default: 1
        description: Номер страницы
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 15
        description: Размер страницы
        in: query
        maximum: 30
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Посты найдены
          schema:
            $ref: '#/definitions/internal_features_posts_transport_http.GetPostByIdDTOResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Поиск постов пользователя
      tags:
      - Posts
  /register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя и возвращает его данные.
      parameters:
      - description: тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_auth_transport_http.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/internal_features_auth_transport_http.RegisterDTOResponse'
        "400":
          description: Неверный запрос или email уже занят
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError'
      summary: Регистрация пользователя
      tags:
      - Auth
swagger: "2.0"